+++ src/App/Restaurants.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+// TODO: Food cards will go here.
+
+/**
+ *
+ */
+const Restaurants = () => (
+    <>
+        <h1>You have selected</h1>
+    </>
+);
+export default Restaurants;
